refactor: 온보딩 스텝의 내부 로직과 제출 로직 분리
