+++ src/app/onboarding/components/steps/PersonalInfoStep.tsx
@@ -1,23 +1,10 @@
 'use client';
 
 import Button from '@/components/buttons/button/Button';
-import RadioButtons from '@/components/buttons/radio-buttons/RadioButtons';
 import Indicator from '@/components/indicator/Indicator';
+import { OnboardingFormValues } from '@/components/onboarding-contents/onboarding.types';
+import PersonalInfoContent from '@/components/onboarding-contents/PersonalInfoContent';
 import { useFormContext } from 'react-hook-form';
-import { OnboardingFormValues } from '../../page';
-import { ERROR_MESSAGES } from '../../constants/formValidation';
-
-const GENDER_OPTIONS = ['여성', '남성'] as const;
-const AGE_OPTIONS = [
-  '10대 이하',
-  '20대',
-  '30대',
-  '40대',
-  '50대',
-  '60대',
-  '70대',
-  '80대 이상',
-] as const;
 
 interface Props {
   handleNextStep: () => void;
@@ -25,52 +12,18 @@ interface Props {
 }
 
 const PersonalInfoStep = ({ handleNextStep, handlePrevStep }: Props) => {
-  const { control, setValue, getValues, trigger } =
-    useFormContext<OnboardingFormValues>();
-  const nickname = getValues('nickname');
+  const { trigger } = useFormContext<OnboardingFormValues>();
 
   const handleCheckStep = async () => {
     const isStepValid = await trigger(['gender', 'age']);
-
     if (isStepValid) {
       handleNextStep();
     }
   };
 
   return (
-    <div className="relative h-full w-full grow">
-      <h2 className="p-28 text-26 font-700 text-grey-900">
-        <span className="text-primary-main">{nickname}</span>님의 <br />
-        성별과 연령대를 알려주세요
-      </h2>
-      <div className="w-full p-28 pb-0">
-        <div className="mb-16 text-16 font-500 text-grey-600-sub">
-          성별을 선택해주세요
-        </div>
-        <RadioButtons
-          values={GENDER_OPTIONS}
-          name="gender"
-          control={control}
-          setValue={setValue}
-          rules={{
-            required: ERROR_MESSAGES.gender.required,
-          }}
-        />
-      </div>
-      <div className="w-full p-28">
-        <div className="mb-16 text-16 font-500 text-grey-600-sub">
-          연령대를 선택해주세요
-        </div>
-        <RadioButtons
-          values={AGE_OPTIONS}
-          name="age"
-          control={control}
-          setValue={setValue}
-          rules={{
-            required: ERROR_MESSAGES.age.required,
-          }}
-        />
-      </div>
+    <>
+      <PersonalInfoContent />
       <div className="absolute bottom-12 flex w-full flex-col items-center bg-white">
         <div className="py-16">
           <Indicator max={4} value={2} />
@@ -88,7 +41,7 @@ const PersonalInfoStep = ({ handleNextStep, handlePrevStep }: Props) => {
           이전으로
         </button>
       </div>
-    </div>
+    </>
   );
 };
