+++ src/app/onboarding/components/steps/ProfileInfoStep.tsx
@@ -1,75 +1,22 @@
 'use client';
 
-import Image from 'next/image';
 import { useFormContext } from 'react-hook-form';
-import CameraIcon from 'public/icons/camera.svg';
 import Indicator from '@/components/indicator/Indicator';
 import Button from '@/components/buttons/button/Button';
-import { ChangeEvent, KeyboardEvent, useEffect, useRef, useState } from 'react';
-import TextInput from '@/components/inputs/text-input/TextInput';
-import { OnboardingFormValues } from '../../page';
-import { ERROR_MESSAGES, REG_EXP } from '../../constants/formValidation';
 import { usePutNickname } from '@/services/users';
-import { toast } from 'react-toastify';
-
-const DEFAULT_PROFILE_SRC = '/icons/default-profile.svg';
-const MAX_FILE_SIZE = 10 * 1024 * 1024; // 10MB
+import { OnboardingFormValues } from '@/components/onboarding-contents/onboarding.types';
+import { ERROR_MESSAGES } from '@/components/onboarding-contents/formValidation.contants';
+import ProfileInfoContent from '@/components/onboarding-contents/ProfileInfoContent';
 
 interface Props {
   handleNextStep: () => void;
 }
 
 const ProfileInfoStep = ({ handleNextStep }: Props) => {
-  const { control, setValue, getValues, trigger, setError, clearErrors } =
+  const { getValues, trigger, setError, clearErrors } =
     useFormContext<OnboardingFormValues>();
 
-  const fileInputRef = useRef<HTMLInputElement>(null);
-  const [imageFile, setImageFile] = useState<File | null>(null);
-  const [imageSrc, setImageSrc] = useState<string | null>(null);
-
-  const showImagePreview = (file: File) => {
-    setImageFile(file);
-    const reader = new FileReader();
-    reader.readAsDataURL(file);
-    reader.onloadend = () => setImageSrc(reader.result as string);
-  };
-
-  const handleFileSelect = (e: ChangeEvent<HTMLInputElement>) => {
-    const file = e.target.files?.[0];
-    if (!file) {
-      return;
-    } else if (file.size > MAX_FILE_SIZE) {
-      toast.error('10MB 이하의 파일만 업로드 할 수 있습니다.');
-      return;
-    }
-    showImagePreview(file);
-  };
-
-  const clearSelectedFile = () => {
-    if (!fileInputRef.current) {
-      return;
-    }
-    setImageFile(null);
-    setImageSrc(null);
-    fileInputRef.current.value = '';
-  };
-
-  useEffect(() => {
-    const savedFile = getValues('profileImage');
-    if (!savedFile) {
-      return;
-    }
-    showImagePreview(savedFile);
-  }, []);
-
-  useEffect(() => {
-    if (!fileInputRef.current) {
-      return;
-    }
-    setValue('profileImage', imageFile);
-  }, [imageFile]);
-
-  const handleNicknameDuplicateError = () => {
+  const setNicknameDuplicateError = () => {
     setError('nickname', {
       type: 'duplicate',
       message: ERROR_MESSAGES.nickname.duplicate,
@@ -83,7 +30,7 @@ const ProfileInfoStep = ({ handleNextStep }: Props) => {
 
   const putNickname = usePutNickname({
     onSuccess: handleValidStep,
-    onError: handleNicknameDuplicateError,
+    onError: setNicknameDuplicateError,
   });
 
   const handleCheckStep = async () => {
@@ -95,74 +42,9 @@ const ProfileInfoStep = ({ handleNextStep }: Props) => {
     putNickname.mutate(newNickname);
   };
 
-  const handleEnter = (e: KeyboardEvent<HTMLInputElement>) => {
-    if (e.key === 'Enter') {
-      handleCheckStep();
-    }
-  };
-
   return (
-    <div className="relative h-full w-full grow">
-      <div className="p-28">
-        <h2 className="pb-[6px] text-26 font-700 text-grey-900">
-          프로필을 입력해주세요
-        </h2>
-        <p className="text-14 font-600 text-grey-500">
-          핸디버스가 어떻게 불러드릴까요?
-        </p>
-      </div>
-      <div className="relative flex h-[286px] w-full flex-col items-center justify-center gap-12 py-28">
-        <div className="relative flex h-200 w-200 items-center justify-center overflow-hidden rounded-full">
-          <Image
-            src={imageSrc || DEFAULT_PROFILE_SRC}
-            alt="프로필 이미지"
-            fill
-            className="object-cover"
-          />
-        </div>
-        <button
-          type="button"
-          onClick={() => fileInputRef.current?.click()}
-          className="absolute bottom-[calc(50%-85px)] right-[calc(50%-85px)] flex h-[34px] w-[34px] items-center justify-center rounded-full bg-black/30"
-        >
-          <CameraIcon />
-          <input
-            ref={fileInputRef}
-            type="file"
-            id="imageUpload"
-            accept="image/*"
-            onChange={handleFileSelect}
-            className="hidden"
-          />
-        </button>
-        {imageFile && (
-          <button
-            type="button"
-            onClick={clearSelectedFile}
-            className="h-[18px] text-14 font-500 text-grey-600-sub underline underline-offset-2"
-          >
-            프로필 사진 지우기
-          </button>
-        )}
-      </div>
-      <div className="p-28">
-        <TextInput
-          name="nickname"
-          control={control}
-          setValue={setValue}
-          onKeyDown={handleEnter}
-          placeholder="영문/한글/숫자 포함 2 ~ 12자"
-          rules={{
-            required: ERROR_MESSAGES.nickname.required,
-            pattern: {
-              value: REG_EXP.nickname,
-              message: ERROR_MESSAGES.nickname.pattern,
-            },
-          }}
-        >
-          닉네임
-        </TextInput>
-      </div>
+    <>
+      <ProfileInfoContent handleSubmit={handleCheckStep} />
       <div className="absolute bottom-[26px] flex w-full flex-col items-center bg-white">
         <div className="py-16">
           <Indicator max={4} value={1} />
@@ -177,7 +59,7 @@ const ProfileInfoStep = ({ handleNextStep }: Props) => {
           </Button>
         </div>
       </div>
-    </div>
+    </>
   );
 };
