+++ src/components/app-bar/AppBar.tsx
@@ -21,7 +21,7 @@ const AppBar = ({ children, handleBack }: Props) => {
 
   return (
     <div className="h-44 w-full px-16 py-12">
-      <button onClick={handleClick} className="flex items-center">
+      <button onClick={handleClick} type="button" className="flex items-center">
         <BackIcon />
         <h1 className="ml-8 text-14 font-600 text-grey-500">{children}</h1>
       </button>
